Organize Lia chatbot into task and exception packages
This commit organizes the code into two packages:
- `task` package for task-related classes: Task, ToDo, Deadline, Event.
- `exception` package for Lia-specific exceptions: LiaException.

This improves code modularity and structure, making it easier to maintain and extend.
