+++ src/main/java/Lia.java
@@ -1,3 +1,8 @@
+import task.Task;
+import task.Event;
+import task.ToDo;
+import task.Deadline;
+import exception.LiaException;
 import java.util.Scanner;
 
 /**
